mm: Handle PTE_MARKER page faults
handle_pte_marker() is the function that will parse and handle all the pte
marker faults.  For PAGEOUT marker, it's as simple as dropping the pte and do
the fault just like a none pte.

The other solution should be that we clear the pte to none pte and retry the
fault, however that'll be slower than handling it right now.

Signed-off-by: Peter Xu <peterx@redhat.com>