mm/userfaultfd: Make uffd read() wait event exclusive
When a new message is generated for an userfaultfd, instead of waking up
all the readers, we can simply only wake up one exclusive reader to process
the event.  Waking up >1 readers for 1 message will definitely be a waste
of resource.

This should make userfaultfd read() O(1) on wakeups.

Use the append-tail version of it, because the order doesn't matter anyway,
so this is even a little bit faster than __add_wait_queue() even though we
don't yet have PRIORITY flag set in any waiters of userfaultfd ctx.

Suggested-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Peter Xu <peterx@redhat.com>