selftests/mm: Replace uffd_read_mutex with a semaphore
Each uffd read threads unlocks the read mutex first, probably just to make
sure the thread is reaching a stage where pthread_cancel() can always work
before the main thread moves on.

However keeping the mutex locked always and unlock in the thread is a bit
hacky.  Replacing it with a semaphore which should be much clearer, where
the main thread will wait() and the thread will just post().  Move it to
uffd-common.* to be reused later.

Signed-off-by: Peter Xu <peterx@redhat.com>