mm/gup: Check pfnmap early for follow_page()
Pfnmaps used to rely on follow_page_pte(), where all the mappings will be
pte_special().

Instead of adding similar checks in all the levels for huge pfnmaps,
follow_page() can fail even earlier because the vma object is on hand and a
decision can be made right at the start.

It's also something gup-slow already does, treat it the same for
follow_page() so that it'll be safe to be invoked on e.g. VM_PFNMAP|VM_IO
where there can be huge pfnmaps.

Note that this should change retval for such old pte pfnmap from -EEXIST to
-EFAULT (which might be a better retval, see follow_pfn_pte()).  But that
shouldn't matter, as the follow_page() callers never checked retvals yet.

Signed-off-by: Peter Xu <peterx@redhat.com>