mm/gup: Check pfnmap early for follow_page()
Pfnmaps used to rely on follow_page_pte(), where all the mappings will be
pte_special(). Huge pfn mappings (pmd+) are coming, so that may not be
enough now.

Instead of adding similar checks in all the levels, follow_page() can
already fail even earlier because the vma object is on hand and a decision
can be made right at the start.

It's also something slow-gup already does, treat it the same for
follow_page() so that it'll be safe to see huge pfn mappings now.

Note that this should change retval for such old pte pfnmap from -EEXIST to
-EFAULT (which should look a better retval comparing to the old, see
follow_pfn_pte()), but that shouldn't matter, as the follow_page() callers
never cared.

Signed-off-by: Peter Xu <peterx@redhat.com>