mm/fork: Accept huge pfnmap entries
Teach the fork code to properly copy pfnmaps for pmd/pud levels.  Pud is
much easier, the write bit needs to be persisted though for writable and
shared pud mappings like PFNMAP ones, otherwise a follow up write in either
parent or child process will trigger a write fault and for such vma it can
get wrongly split (which can cause the entry be zapped for a PFNMAP).

Do the same for pmd level.  Note that for code simplicity the special bit
is only checked lockless. Recheck is not needed in this case as the entry
can't change under mmap write lock.

Signed-off-by: Peter Xu <peterx@redhat.com>