mm: mempolicy: Refactor out policy_node_nodemask()
This was refactored out of huge_node().

huge_node()'s interpretation of vma for order assumes the
hugetlb-specific storage of the hstate information in the
inode. policy_node_nodemask() does not assume that, and can be used
more generically.

This refactoring also enforces that nid default to the current node
id, which was not previously enforced.

alloc_pages_mpol_noprof() is the last remaining direct user of
policy_nodemask(). All its callers begin with nid being the current
node id as well. More refactoring is required for to simplify that.

Signed-off-by: Ackerley Tng <ackerleytng@google.com>
Reviewed-by: Gregory Price <gourry@gourry.net>
Link: https://lore.kernel.org/r/9831cfcc77e325e48ec3674c3a518bda76e78df5.1726009989.git.ackerleytng@google.com
Signed-off-by: Peter Xu <peterx@redhat.com>