mm: Introduce vmf_insert_pmd()
Introduce vmf_insert_pmd(), which is almost what do_set_pmd() does, except
that it decouples any THP related operations.  It is to be reused later to
install non-THP pmd entries.

The name vmf_insert_pmd() came from vmf_insert_pfn_pmd(), it should work
almost like that except that it works with a folio rather than pfnmaps.
And this is also a chance we make the parameter to be a folio not page,
which keeps do_set_pmd() as before, which supports tail pages.

Signed-off-by: Peter Xu <peterx@redhat.com>