mm: Allow vma to opt-in on report supported orders
So far, most huge page fault supports are hard-coded so far in
thp_vma_allowable_order(), even if some of them may not be THPs (DAX,
pfnmaps).

This patch provides a generic way for a VMA to report huge mapping support
for drivers, especially this allows that to overwrite / ignore THP
settings.

For now, only introduce this framework, make it prepared to be used in new
drivers.  Logically it can also apply to existing drivers. Take hugetlbfs
as an example, if such will be supported in generic mm fault path one day,
hugetlbfs can report on corresponding page size supported in the hook fn.

Signed-off-by: Peter Xu <peterx@redhat.com>