mm: Allow zap huge puds for normal file mappings
We're going to introduce the first pud mapping for generic file.  Loose the
protection for zap.

The pud split path needs similar care.  So far split a huge pud always
means clearing the entry.  It'll not be the same when there's file pud
folio to be mapped which uses genuine refcount / mapcount accountings.  We
need to do proper rmap accounting and update the statistics.

Signed-off-by: Peter Xu <peterx@redhat.com>