mm: Change retval for vmf_insert_pmd() when entry existed
When the core mm wants to inject a huge pmd entry, it used to return
VM_FAULT_FALLBACK when it sees the entry was populated by another thread.

This may not be correct, and it's prone of issues.

The problem is VM_FAULT_FALLBACK normally implies "let's keep walking the
next level of pgtable".  When the populated entry is also a huge pmd (which
is very likely, because obviously when it can race in vmf_insert_pmd(), it
means pmd entry is eligible at this address), suggesting to walk the next
level is almost wrong.  The correct way is to retry the fault, then we'll
see whether it's a huge pmd or a pte pgtable that has been installed.

Here return VM_FAULT_NOPAGE instead.  It's a slight abuse here, because it
doesn't resolve the page fault itself in the current fault(), however
accurate from at least code level to request for a retry.

This should have no functional change for most existing callers, except
this code path:

  finish_fault -> do_set_pmd -> vmf_insert_pmd

The functional change here is hopefully wanted, because otherwise
finish_fault() will continue with pmd_install() or pte_alloc() even if a
huge pmd populated, which looks wrong.

This also paves way for the future users of this function.

Signed-off-by: Peter Xu <peterx@redhat.com>